Remove all remaining references to OptimizerContext
