+++ dawn/src/dawn/IIR/MultiStage.h
@@ -27,7 +27,6 @@
 #include <vector>
 
 namespace dawn {
-class OptimizerContext;
 namespace iir {
 
 class Stencil;
