+++ dawn/src/dawn/Optimizer/Pass.h
@@ -21,7 +21,7 @@
 #include <vector>
 
 namespace dawn {
-class OptimizerContext;
+
 namespace iir {
 class StencilInstantiation;
 }
