+++ dawn/src/dawn/Optimizer/PassFixVersionedInputFields.cpp
@@ -21,7 +21,6 @@
 #include "dawn/IIR/IIRNodeIterator.h"
 #include "dawn/IIR/NodeUpdateType.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Support/Logger.h"
 
 #include <memory>
