+++ dawn/src/dawn/Optimizer/PassInlining.cpp
@@ -20,7 +20,6 @@
 #include "dawn/IIR/IIRNodeIterator.h"
 #include "dawn/IIR/InstantiationHelper.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Support/Logger.h"
 #include "dawn/Support/STLExtras.h"
