+++ dawn/src/dawn/Optimizer/PassIntervalPartitioning.cpp
@@ -16,7 +16,6 @@
 #include "dawn/IIR/IIRNodeIterator.h"
 #include "dawn/IIR/Stencil.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 
 namespace dawn {
