+++ dawn/src/dawn/Optimizer/PassSetCaches.cpp
@@ -16,7 +16,6 @@
 #include "dawn/IIR/Cache.h"
 #include "dawn/IIR/IntervalAlgorithms.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Support/Logger.h"
 #include "dawn/Support/Unreachable.h"
 #include <optional>
