+++ dawn/src/dawn/Optimizer/PassSetStageName.cpp
@@ -14,7 +14,6 @@
 
 #include "dawn/Optimizer/PassSetStageName.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 
 namespace dawn {
