+++ dawn/src/dawn/Optimizer/PassTemporaryFirstAccess.cpp
@@ -16,7 +16,6 @@
 #include "dawn/IIR/ASTVisitor.h"
 #include "dawn/IIR/IIRNodeIterator.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Support/Exception.h"
 #include "dawn/Support/IndexRange.h"
 #include "dawn/Support/Logger.h"
