+++ dawn/src/dawn/Optimizer/PassTemporaryToStencilFunction.cpp
@@ -22,7 +22,6 @@
 #include "dawn/IIR/IIRNodeIterator.h"
 #include "dawn/IIR/Stencil.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Optimizer/StatementMapper.h"
 #include "dawn/Optimizer/TemporaryHandling.h"
 #include "dawn/SIR/AST.h"
