+++ dawn/src/dawn/Optimizer/PassTemporaryType.cpp
@@ -19,7 +19,6 @@
 #include "dawn/IIR/NodeUpdateType.h"
 #include "dawn/IIR/Stencil.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Optimizer/TemporaryHandling.h"
 #include "dawn/Support/Logger.h"
 #include <memory>
