+++ dawn/src/dawn/Unittest/IIRBuilder.cpp
@@ -28,7 +28,7 @@
 
 #include "dawn/IIR/AccessComputation.h"
 #include "dawn/IIR/InstantiationHelper.h"
-#include "dawn/Optimizer/OptimizerContext.h"
+#include "dawn/Optimizer/Lowering.h"
 #include "dawn/Validator/GridTypeChecker.h"
 #include "dawn/Validator/UnstructuredDimensionChecker.h"
 #include "dawn/Validator/WeightChecker.h"
