+++ dawn/test/unit-test/dawn/Optimizer/TestPassCaching.cpp
@@ -15,7 +15,6 @@
 #include "dawn/CodeGen/CXXNaive/CXXNaiveCodeGen.h"
 #include "dawn/CodeGen/CodeGen.h"
 #include "dawn/CodeGen/Cuda/CudaCodeGen.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Optimizer/PassSetCaches.h"
 #include "dawn/Unittest/IIRBuilder.h"
