+++ dawn/test/unit-test/dawn/Optimizer/TestPassFieldVersioning.cpp
@@ -13,7 +13,6 @@
 //===------------------------------------------------------------------------------------------===//
 #include "dawn/IIR/IIR.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Optimizer/PassFieldVersioning.h"
 #include "dawn/Serialization/IIRSerializer.h"
