+++ dawn/test/unit-test/dawn/Optimizer/TestPassRemoveScalars.cpp
@@ -11,7 +11,6 @@
 //  See LICENSE.txt for details.
 //
 //===------------------------------------------------------------------------------------------===//
-#include "dawn/Optimizer/OptimizerContext.h"
 #include "dawn/Optimizer/PassRemoveScalars.h"
 #include "dawn/Support/Logger.h"
 #include "dawn/Unittest/ASTConstructionAliases.h"
@@ -118,10 +117,6 @@ TEST(TestRemoveScalars, test_unstructured_scalar_02) {
   int varAID = metadata.getAccessIDFromName("varA");
   int varBID = metadata.getAccessIDFromName("varB");
 
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
-
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
 
@@ -166,10 +161,6 @@ TEST(TestRemoveScalars, test_cartesian_scalar_01) {
   auto& metadata = stencil->getMetaData();
   int varAID = metadata.getAccessIDFromName("varA");
   int varBID = metadata.getAccessIDFromName("varB");
-
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
 
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
@@ -216,10 +207,6 @@ TEST(TestRemoveScalars, test_cartesian_scalar_02) {
   int varAID = metadata.getAccessIDFromName("varA");
   int varBID = metadata.getAccessIDFromName("varB");
 
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
-
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
 
@@ -265,10 +252,6 @@ TEST(TestRemoveScalars, test_global_01) {
   auto& metadata = stencil->getMetaData();
   int varAID = metadata.getAccessIDFromName("varA");
 
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
-
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
 
@@ -312,10 +295,6 @@ TEST(TestRemoveScalars, test_if_01) {
 
   auto& metadata = stencil->getMetaData();
   int varAID = metadata.getAccessIDFromName("varA");
-
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
 
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
@@ -364,10 +343,6 @@ TEST(TestRemoveScalars, test_if_02) {
 
   auto& metadata = stencil->getMetaData();
   int varAID = metadata.getAccessIDFromName("varA");
-
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
 
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
@@ -421,10 +396,6 @@ TEST(TestRemoveScalars, test_else_01) {
   auto& metadata = stencil->getMetaData();
   int varAID = metadata.getAccessIDFromName("varA");
 
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
-
   PassRemoveScalars passRemoveScalars;
   passRemoveScalars.run(stencil);
 
@@ -472,10 +443,6 @@ TEST(TestRemoveScalars, warn_compound_assignments) {
                                      b.stmt(b.assignExpr(b.at(varA), b.lit(3.0), Op::multiply)),
                                      b.stmt(b.assignExpr(b.at(f_e), b.at(varA))))))));
 
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
-
   std::ostringstream output;
   dawn::log::info.stream(output);
   dawn::log::setVerbosity(dawn::log::Level::All);
@@ -506,10 +473,6 @@ TEST(TestRemoveScalars, warn_increment) {
           b.stage(b.doMethod(dawn::sir::Interval::Start, dawn::sir::Interval::End,
                              b.declareVar(varA), b.stmt(b.unaryExpr(b.at(varA), Op::increment)),
                              b.stmt(b.assignExpr(b.at(f_e), b.at(varA))))))));
-
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
 
   std::ostringstream output;
   dawn::log::info.stream(output);
@@ -545,10 +508,6 @@ TEST(TestRemoveScalars, warn_condition_adimensional_01) {
                                     b.block(b.stmt(b.assignExpr(b.at(varA), b.lit(4.0))))),
                            b.stmt(b.assignExpr(b.at(f_e), b.at(varA))))))));
 
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
-
   std::ostringstream output;
   dawn::log::info.stream(output);
   dawn::log::setVerbosity(dawn::log::Level::All);
@@ -582,10 +541,6 @@ TEST(TestRemoveScalars, warn_condition_adimensional_02) {
                            b.ifStmt(b.binaryExpr(b.lit(1.0), b.lit(1.0), Op::minus),
                                     b.block(b.stmt(b.assignExpr(b.at(varA), b.lit(4.0))))),
                            b.stmt(b.assignExpr(b.at(f_e), b.at(varA))))))));
-
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
 
   std::ostringstream output;
   dawn::log::info.stream(output);
@@ -626,10 +581,6 @@ TEST(TestRemoveScalars, warn_condition_adimensional_03) {
   std::ostringstream output;
   dawn::log::info.stream(output);
   dawn::log::setVerbosity(dawn::log::Level::All);
-
-  OptimizerContext::OptimizerContextOptions optimizerOptions;
-  OptimizerContext optimizer(optimizerOptions,
-                             std::make_shared<dawn::SIR>(ast::GridType::Unstructured));
 
   // run single pass (PassRemoveScalars) and expect info in output
   PassRemoveScalars passRemoveScalars;
